Create NameSpoofer
Hi everyone this is my first addon hope you all like it.
NameSpoofer is an addon that changes your in-game name as many times as you want without a problem. It is useful if you want to hide your real name. It is also useful if you play with cheats/scripts when you stream. This addon works for Warmane but can work for everyone else.